+++ include/sched.h
@@ -6,28 +6,30 @@
 
 #ifndef __ASSEMBLER__
 
+#include <stdint.h>
+
 struct cpu_context {
-    unsigned long x19;
-    unsigned long x20;
-    unsigned long x21;
-    unsigned long x22;
-    unsigned long x23;
-    unsigned long x24;
-    unsigned long x25;
-    unsigned long x26;
-    unsigned long x27;
-    unsigned long x28;
-    unsigned long fp;
-    unsigned long sp;
-    unsigned long pc;
+    uint64_t x19;
+    uint64_t x20;
+    uint64_t x21;
+    uint64_t x22;
+    uint64_t x23;
+    uint64_t x24;
+    uint64_t x25;
+    uint64_t x26;
+    uint64_t x27;
+    uint64_t x28;
+    uint64_t fp;
+    uint64_t sp;
+    uint64_t pc;
 };
 
 struct task_struct {
     struct cpu_context cpu_context;
-    long state;	
-    long counter;
-    long priority;
-    long preempt_count;
+    int32_t state;	
+    int32_t counter;
+    int32_t priority;
+    int32_t preempt_count;
 };
 
 void sched_init(void);
@@ -36,7 +38,7 @@ void sched_preempt_enable(void);
 void sched_schedule(void);
 void sched_timer_tick(void);
 void sched_switch_to(struct task_struct *next);
-long sched_get_current_priority(void);
-int sched_add_fork(struct task_struct *p);
+int32_t sched_get_current_priority(void);
+int8_t sched_add_fork(struct task_struct *p);
 
 #endif
